requirements.txt
dm-haiku
einops
jax
jaxlib
numpy
optax
Pillow
